+++ src/cryptography/hazmat/primitives/hashes.py
@@ -103,11 +103,23 @@ class SHA1(HashAlgorithm):
     digest_size = 20
     block_size = 64
 
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, SHA1):
+            return NotImplemented
+
+        return True
+
 
 class SHA512_224(HashAlgorithm):  # noqa: N801
     name = "sha512-224"
     digest_size = 28
     block_size = 128
+
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, SHA512_224):
+            return NotImplemented
+
+        return True
 
 
 class SHA512_256(HashAlgorithm):  # noqa: N801
@@ -115,11 +127,23 @@ class SHA512_256(HashAlgorithm):  # noqa: N801
     digest_size = 32
     block_size = 128
 
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, SHA512_256):
+            return NotImplemented
+
+        return True
+
 
 class SHA224(HashAlgorithm):
     name = "sha224"
     digest_size = 28
     block_size = 64
+
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, SHA224):
+            return NotImplemented
+
+        return True
 
 
 class SHA256(HashAlgorithm):
@@ -127,11 +151,23 @@ class SHA256(HashAlgorithm):
     digest_size = 32
     block_size = 64
 
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, SHA256):
+            return NotImplemented
+
+        return True
+
 
 class SHA384(HashAlgorithm):
     name = "sha384"
     digest_size = 48
     block_size = 128
+
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, SHA384):
+            return NotImplemented
+
+        return True
 
 
 class SHA512(HashAlgorithm):
@@ -139,11 +175,23 @@ class SHA512(HashAlgorithm):
     digest_size = 64
     block_size = 128
 
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, SHA512):
+            return NotImplemented
+
+        return True
+
 
 class SHA3_224(HashAlgorithm):  # noqa: N801
     name = "sha3-224"
     digest_size = 28
     block_size = None
+
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, SHA3_224):
+            return NotImplemented
+
+        return True
 
 
 class SHA3_256(HashAlgorithm):  # noqa: N801
@@ -151,17 +199,35 @@ class SHA3_256(HashAlgorithm):  # noqa: N801
     digest_size = 32
     block_size = None
 
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, SHA3_256):
+            return NotImplemented
+
+        return True
+
 
 class SHA3_384(HashAlgorithm):  # noqa: N801
     name = "sha3-384"
     digest_size = 48
     block_size = None
 
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, SHA3_384):
+            return NotImplemented
+
+        return True
+
 
 class SHA3_512(HashAlgorithm):  # noqa: N801
     name = "sha3-512"
     digest_size = 64
     block_size = None
+
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, SHA3_512):
+            return NotImplemented
+
+        return True
 
 
 class SHAKE128(HashAlgorithm, ExtendableOutputFunction):
@@ -176,6 +242,12 @@ class SHAKE128(HashAlgorithm, ExtendableOutputFunction):
             raise ValueError("digest_size must be a positive integer")
 
         self._digest_size = digest_size
+
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, SHAKE128):
+            return NotImplemented
+
+        return self._digest_size == other._digest_size
 
     @property
     def digest_size(self) -> int:
@@ -195,6 +267,12 @@ class SHAKE256(HashAlgorithm, ExtendableOutputFunction):
 
         self._digest_size = digest_size
 
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, SHAKE256):
+            return NotImplemented
+
+        return self._digest_size == other._digest_size
+
     @property
     def digest_size(self) -> int:
         return self._digest_size
@@ -204,6 +282,12 @@ class MD5(HashAlgorithm):
     name = "md5"
     digest_size = 16
     block_size = 64
+
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, MD5):
+            return NotImplemented
+
+        return True
 
 
 class BLAKE2b(HashAlgorithm):
@@ -217,6 +301,12 @@ class BLAKE2b(HashAlgorithm):
             raise ValueError("Digest size must be 64")
 
         self._digest_size = digest_size
+
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, BLAKE2b):
+            return NotImplemented
+
+        return self._digest_size == other._digest_size
 
     @property
     def digest_size(self) -> int:
@@ -235,6 +325,12 @@ class BLAKE2s(HashAlgorithm):
 
         self._digest_size = digest_size
 
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, BLAKE2s):
+            return NotImplemented
+
+        return self._digest_size == other._digest_size
+
     @property
     def digest_size(self) -> int:
         return self._digest_size
@@ -244,3 +340,9 @@ class SM3(HashAlgorithm):
     name = "sm3"
     digest_size = 32
     block_size = 64
+
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, SM3):
+            return NotImplemented
+
+        return True